% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemVI.R
\name{chemVI}
\alias{chemVI}
\title{Chemical Variable Importance for Floating Percentile Model Benchmarks}
\usage{
chemVI(data, paramList, ...)
}
\arguments{
\item{data}{data.frame containing, at a minimum, chemical concentrations as columns and a logical \code{Hit} column classifying toxicity}

\item{paramList}{character vector naming columns of \code{data} containing concentrations}

\item{...}{additional arguments passed to \code{chemSig}, \code{chemSigSelect}, and \code{FPM}}
}
\value{
data.frame with 2 columns
}
\description{
Generate statistics describing the relative importance of chemicals among benchmarks generated by \code{FPM}
}
\details{
The purpose of \code{chemVI} is to inform the user about the relative influence of each chemical over the sediment quality benchmarks generated by \code{FPM}.
Three statistics are generated: \code{chemDensity}, \code{MADP}, \code{dOR}, \code{dFM}, and \code{dMCC}. The \code{chemDensity} statistic (which is also generated by \code{FPM})
describes how little a particular chemical's value increased within the floating percentile model algorithm.
Low \code{chemDensity} (close to 0) means that the value was able to increase substantially within the algorithm without triggering one or more of the criteria for
stopping the algorithm (see \code{?FPM}), whereas high \code{chemDensity} (close to 1) indicates the final benchmark for that chemical did not float (increase)
much before being locked in. In other words, low \code{chemDensity} might be interpreted as relatively low importance. We caution against using this
metric in isolation, as it is the more difficult to interpret of the three.
The \code{MADP} statistic (or mean absolute difference percent) is calculated by sequentially dropping each chemical from consideration, recalculating the benchmarks
for the remaining chemicals, and then determining how much each benchmark changed (as a percent of the original value). Thus, the \code{MADP}
is a measure of a chemical's influence over other benchmarks. The \code{dOR} statistic is the difference between the overall reliability
of benchmarks with all chemicals versus without each chemical. \code{dFM} and \code{dMCC} are similar to the \code{dOR} statistic, but for the Fowlkes-Mallows Index
and Matthew's Correlation Coefficient. In any case, larger positive values indicate a greater impact of a chemical
on the overall predictive performance of floating percentile model benchmarks. Small values (close to 0) indicate low influence. Larger negative values indicate that
the chemical actually adversely impacts toxicity predictions. If there are chemicals with negative values, consider reevaluting the data without the associated chemical
or using \code{optimFPM} or \code{cvFPM} to optimize the overall reliability prior to running \code{FPM} and \code{chemVI}.
}
\examples{
paramList = c("Cd", "Cu", "Fe", "Mn", "Ni", "Pb", "Zn")
chemVI(h.tristate, paramList, testType = "np")
chemVI(h.tristate, paramList, testType = "p")
}
\seealso{
chemSig, chemSigSelect, optimFPM, cvFPM, FPM
}
